pyproject.toml
[tool.poetry]
name = "pharmacies_saas2"
version = "0.1.0"
description = "Django SaaS backend"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
python-decouple = "^3.8"
dj-database-url = "^3.1.2"
django-extensions = "^4.1"
djangorestframework = "^3.17.1"
djangorestframework-simplejwt = "^5.5.1"
sqlparse = "^0.5.5"
gunicorn = "^25.3.0"
django-cors-headers = "^4.9.0"
whitenoise = "^6.12.0"
psycopg = {extras = ["binary"], version = "^3.3.3"}
requests = "^2.33.1"
uvicorn = "^0.44.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pluggy = "^1.6.0"
packaging = "^26.0"
